Fixed a bug where screenshots from a later test overwrite ones from an earlier test: if test1.feature and test2.feature both take screenshots on line 2, then test2.feature's screenshot overwrites test1.feature's screenshot because the filename is the same.
I've fixed this by having a test artifacts directory set. A directory is
created for every test .feature file and all the screenshots are filed
away in there. The test artifacts directory is set by setting the
TEST_ARTIFACTS_DIR environment variable. If it is not set, this defaults
to "results"