Added IsCurrentActivityFocused action.
This is useful to run tests across different applications (for example,
your application launching the browser). Since the instrumentation can
only work with the application under test, then all we can assert is that
the application has lost the focus.

This is to cover Robotium's limitation around testing across 2 different applications.