Fixed #135
The problem seems to be a simple name collision between Calabash and Cucumber
for the use of the keyword "background".

Renaming Calabash (not yet implemented) "background" operation to app_to_background
solves the issue.

Even if this in theory breaks backwards compatibility, since the operation hadn't
been implemented yet it's going to hardly bother anybody.