Adding set_locale to the calabash actions
The set_locale action sets the current app configuration to use the
language and locale codes specified in the locale cucumber steps. For
example: I change the language code to "fr" and region code to
"rCA" to change the app to Canadian French. After changing the language,
it restarts the activity (so that the activity is forced to restart its
lifecycle and load the new language).