Fixed regexp for l10nkey definitions....
In continuing to work with these in my own tests, I see that the regexp only works with numeric l10n keys.  It would be more useful if it worked with non-quote characters as in the definition of:

`I wait for the translated "key" l10nkey to appear`

Fixing the docs here, will also submit a pull request for the code.