Rake: use server build script and don't stage server sources
> use server/bin/build.sh

This is an attempt to unify the build process in the server repo and the
build process in the gem.  For example, in the server the
targetSdkVersion has been set to 22, but the rake build task used 19.

> staging

There appears to be no reason to stage the sources to a tmp directory
because the server/bin/build.sh script already performs a staging step.